requirements.txt
asgiref==3.7.2
Django==4.1.10
django-cors-headers==4.2.0
djangorestframework==3.14.0
dnspython==2.4.2
psycopg2-binary==2.9.7
pytz==2023.3
sqlparse==0.2.4
